Add Book Description Feature
This feature is for a Writer and an Investor.
And additional button and page is added in episode that can only seen by an Investor. The edit/add feature is available for a Writer through Edit Book

Also, in addition, I removed a mysterious 's' character from the top of bookfinaldraft.jsp